Major overhaul
Based on the new research by Aaron Bell, the glyphList revised.

Following the remarks by Aaron Bell on the first text proof, anchors and comb accents were repositioned and checked. Including the (Central) European com accents.

Kerning was added for the (tabular) seven, as it is being used as glotal mark in the indigenous languages.